12-04-23

+++ tema3/bd/read.php
@@ -1,4 +1,5 @@
 <?php
+include('verificar.php');
 include('conexion.php');
 $sql = "SELECT id,nombres,apellidos,celular FROM personas";
 $resultado = $con->query($sql);
@@ -17,9 +18,12 @@ if ($resultado->num_rows > 0) {
                 <td><?php echo $row['nombres'] ?> </td>
                 <td><?php echo $row['apellidos'] ?></td>
                 <td><?php echo $row['celular'] ?></td>
-                <td><a href="form_update.php?id=<?php echo $row['id']; ?>">Editar</a>
-                <a href="delete.php?id=<?php echo $row['id']; ?>">Eliminar</a>
-            </td>
+                <td>
+                    <?php if ($_SESSION['nivel'] == 1) { ?>
+                        <a href="form_update.php?id=<?php echo $row['id']; ?>">Editar</a>
+                        <a href="delete.php?id=<?php echo $row['id']; ?>">Eliminar</a>
+                    <?php } ?>
+                </td>
             </tr>
         <?php } ?>
     </table>
@@ -29,8 +33,11 @@ if ($resultado->num_rows > 0) {
 } else {
     echo "la tabla no tiene datos que mostrar";
 }
-
+if ($_SESSION['nivel']==1) {
+?>  
+<a href="form_create.php">Insertar</a>
+<?php
+}
 $con->close();
 ?>
 
-<meta http-equiv="refresh" content="3; url=read.php" />
